pyproject.toml
[tool.poetry]
name = "feature-extractor"
version = "0.1.0"
description = "A feature extractor for Lookyloo webpage captures"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
lingua-language-detector = "^2.0.2"
tldextract = "^5.1.2"
pylookyloo = "^1.25.0"
bs4 = "^0.0.2"
dnspython = "^2.6.1"
lxml = "^5.2.2"
spacy = "^3.7.5"
wordcloud = "^1.9.3"
pandas = "^2.2.2"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
 